requirements_render.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
langgraph>=0.0.40
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
